requirements.txt
gradio
transformers
torch
Pillow
deep-translator
openpyxl
pandas
evaluate
sacrebleu